[WIP] Handle new block storage in 20w17a
This is a (very-WIP) first attempt to handle the restructuring of
BlockStates in chunks written by Minecraft Snapshot 20w17a.

My servers map has a combination of new and old blocks, so this code
should handle both (with the exception of 12-bit bits-per-block in
chunks, since these pack to the same number of Blocks as 11; I have
some ideas on how to detect this case but this is not implemented yet)

So far its a simple first implementation, when I've tested more I'll
look at performance enhancing it.